backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-generativeai
google-api-python-client
google-auth
PyJWT
requests
